Refactor servlets to use util methods; move StoreRecipe to History and SignUp to Account servlets
