+++ backend/src/main/java/com/google/sps/servlets/TextToSpeechServlet.java
@@ -24,7 +24,6 @@ import com.google.cloud.texttospeech.v1.VoiceSelectionParams;
 import com.google.gson.Gson;
 import com.google.protobuf.ByteString;
 import java.io.IOException;
-import java.util.Arrays;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
