+++ frontend/src/components/individual-recipe/Tutorial.js
@@ -173,7 +173,7 @@ class Tutorial extends Component {
   }
 
   finishCooking(recipe) {
-    const request = new Request("/api/store-recipe", {
+    const request = new Request("/api/history", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
